Fill the reporting forms
Mostly could be processed using the README.md
with a reference to the current branch structure.
A propos, the PRISMA-S checklist could be filled
completely at this stage.

In fact, what is left for the individual RESCUERs to report
apart from the data already indicated in the README.md,
are the methods and results of data charting, analysis,
and synthesis, which are rather specific to the context.