Change some date tags & spell all except tags
Changed MHDA dates for PTs/SBs to DCOM
because as I figured PTs are not the same thing
as MeSH, MHDA is not the right concept;
so using DCOM/CDAT is the only way we can
get around with that.

Also, attached a CRDT to all text chunks directly
to be sure this is processed correctly.
This is the most reliable first-entry tag:
https://www.nlm.nih.gov/pubs/techbull/nd08/nd08_pm_new_date_field.html

Moreover, for accuracy, I changed the dummy 1600/01/01[crdt]
to the actual 1865/01/01[crdt], which I identified
via running test searches.

TWs got a DCOM date because they crawl through
MeSH terms & PTs as well.

Also, decided to put in all quotation marks,
change all to lowercase, and spell all the booleans,
so that the query looked like the one from the Search
Details from the beginning.
Like, WYSIWYG.
I still didn’t spell tags because that’s irrelevant
to the query syntax.