Find an optimal TW-DCOM solution & rerun search
👌

What is actually very good, it’s that the record count
almost reached that of the live version
(less than 2k records fewer).

It still loses some textual info that could be queried
since CRDT (like the author or institution list),
but those losses are likely small enough;
and to stress: those losses are not actual losses
but rather delays instead.

Thus, to sum up, this solution real does seem
to be the optimal one.