Commit posts-edit eventually (Sep 2 only)
Considered it previously and decided not to:

https://github.com/p1m-ortho/qs-global-ortho-search-queries/commit/981b05210b740974ca518f8c8b3f2fee93b73ef7

But now, figured it would be super cool
if anyone could fork the or PR these.

And also, having a backup of these
proves useful after implementing changes
(because I tend to overwrite them).

These were built with latest GMP 2.2.3
(BTW, need to add a note with GMP version
in the editables).

Also, contains the preview of r6d2. 😊