Announce appraisal-speeding experiment
In full, the speeding-up experiment consists of 3 parts:

* Rearrange footer for editables (see parent commit)
* Reorganize specialty tagging (this commit)
* Consider changes to screening and titling procedures

As to the last one, consider this:

* Specialty tagging cannot, naturally, be dropped
as it’s central to Zheln’s mission to sort records by specialty ASAP.
* Record screening against Shojania & Bero criteria
could, arguably, be automated using natural language processing.
However, at the moment, I lack expertise to do this
reliably enough and don’t have resources to hire.
* Writing titles for records, too, could arguably be dropped
or automated;
however, I believe this is also critical for good-looking
Zheln records and cannot be automated
reliably enough with my current skill.

Hence the current structure of this speeding-up exercise.