Appraise two more Sep 27 & report new note
Wanted to note I’ve been using `60yo+`
or `6–12yo`-like patterns in titles to denote ages,
that are prevalent; so consistent use of these
throughout is warranted.

Also, I’m starting to figure I’m not selecting
COVID-19 studies for immediate appraisal, as a rule.
This is because there already are plenty
of great resources, covid-nma.com, covidex.ai,
WHO and Cochrane resources, just to name a few.
And also, I don’t like the idea of shifting global
Zheln focus to covid.