Bring a little more clarity to methods section
At first, figured it’d be better to put stages
before the steps; however, eventually, decided
that stages is secondary information.

Also, I was worried the primary appraisal section
didn’t get referenced from the ten steps section,
but now I think it’s okay as this information
is not important for everyone and is less granular
than the steps.

However, this information does shed light
on how the actual appraisal process looks like,
so it is important. I am now thinking to report
primary and full appraisal protocols separately,
so this is going to be taken into account.