Start drafting specialty groups
This is how they are gonna look like.

The groups contain not only
specialties for subspecialties
but also synonymous subspecialties
that link to one and the same page,
which have been an issue of tagging
effort duplication for a while now.

Will have to complete the groups
and also modify General Makeposti
and footer-edit to use and show
synonymous specialties as one item.

Ref. https://github.com/drzhelnov/zheln.github.io/issues/147