Appraise 13 of 417 Dec 9 records total + 1st dupe
Appraised up to 1st selectee for full appraisal.

* This is the first case where I deal
with record duplication.
* I stated in the protocol that duplicates
will not be resolved because records
are appraised within one day, so this is how
I’m going to deal with it: let us think that
the duplicate record is an independent
record but just note in its title that it’s a dupe.
* Also, duplication will be noted in flow diagram.
* In the appraisal log, this is now noticed in italics.
Also, will put all Full Text Wanted appraisals
in bold
in the next commit.

Issue ref. https://github.com/drzhelnov/zheln.github.io/issues/163