Upd appraisal log 2024-06-09 thru 2024-10-27
This period has been covered by ZaaS AI pipeline (yet to be
released publicly), which runs the searches in code, so no separate
files exist anymore to be published in the "ssb-reproducible-8ac8866"
directory. Thus, I am attaching the code version that is now in use.
It is the same 8ac8866 version of the RSSB, the one with the EDAT fix
introduced in:

8ac886686c3b481e6fe20d622d128687087cf4f5

Also, the downloaded records are now in pmid format, not text format,
because ZaaS AI pipeline relies on OpenAlex metadata.

Here are the dates when the attached search updates were conducted:

search_update_date	start_date	end_date
2024-07-28	2024-06-09	2024-06-30
2024-07-28	2024-07-01	2024-07-27
2024-09-15	2024-07-28	2024-08-10
2024-09-15	2024-08-11	2024-09-14
2024-10-28	2024-09-15	2024-09-30

As always, all dates are in Anywhere on Earth (UTC-12) time zone.

I will continue to update the appraisal log from time to time,
as well as publish the pmid list dumps and report any changes
to the search strategy, to ensure the continuity. Of note, the numbers
in the Total column of appraisal log are now simply the output
from wc -l command for the respective pmid_list file.

Thank you for staying with Zheln!